GPII-545: Adding the new GPII usb user listener code
The new GPII usb user listener is now a standalone application.
It covers all the workflow that we need by itself, from detecting the
device to trigger the login/logout processes.

Since the user listener now is standalone, it should be installed in any
well-known 'bin' folder by the system, this is, a 'bin' folder which is
included into $PATH environment variable.

Also, we are providing a '.desktop' file for the users to run it easily
in the same way they launch any other application from the system.