GPII-288: Copy Orca settings for current token(s)
This is part of the Linux scripts for writing the local settings to a log file. The required JavaScript part for Node.js is in a pull request for the pilots2 branch (a branch of the universal repo): https://github.com/GPII/universal/pull/239. (It's also my first Python script. Show no mercy :-P )