Create LogLocalSettings.sh
This is part of the Linux scripts for writing the local settings to a log file. The required JavaScript part for Node.js is in a pull request for the pilots2 branch (a branch of the universal repo): https://github.com/GPII/universal/pull/239. The required Python script is in the other commit for this pull request. (It's also my first bash script. Show no mercy :-P )