GPII-442:  Detect running solutions.
Firmed up the c++ code for the 'nodeprocesse' add-on:  simply a
bridge to gnome/linux library to generate a list of processes and
information about each.

Added numerous unit tests for the above, where most of the tests
use an object that provides an API for processes.

Next step:  carve out the relevant functions into a non-test object
that provides the same API.