GPII-1081: Fedora 21 PackageKit node add-on unit test failure.
As per v1.0.3 of PackageKit, added pkTests.availableNotInstalled()
to check if the package list returned using the "~installed"
filter includes an already installed package, and ignores that
package if found.  Used by the "search" and "get" tests.

This is backward compatible with the way the previous version of
PackageKit worked (v0.8.17), so the tests continue to work with
that version as well.